requirements.txt
# Backend (Django) Dependencies
Django==5.0.1
djangorestframework==3.14.0
psycopg2==2.9.9   # PostgreSQL adapter, you may remove if using SQLite
django-cors-headers==4.3.1
djangorestframework-simplejwt== 5.3.1



